A2) zweites Tetraeder, beide rotieren, Vorbereitung "tether Mode", fix Pfeiltasten immer noch benutzbar
nur Pfeiltasten verändern Ort des Zentrums -> Tetraeder eiern

+++ A1_Versuch1c/Versuch1c.cpp
@@ -225,7 +225,7 @@ void setupGeometry()
 	}
 }
 
-std::vector<uint16_t> indices = {
+std::vector<uint32_t> indices = {
 	0, 1, 2,
 	0, 2, 3,
 	0, 3, 4,
